refactor: linux 单独处理偏好设置窗口的 `skipTaskbar` 参数
