+++ src-tauri/src/core/setup/windows.rs
